Move out parser, convert config to context
