+++ src/context/mod.rs
@@ -1,5 +1,4 @@
 pub mod common;
 pub mod model;
-pub mod parser;
 pub mod pipeline;
 pub mod transform;
