+++ src/lib.rs
@@ -1,5 +1,6 @@
-pub mod config;
+pub mod context;
 pub mod model;
+pub mod parser;
 pub mod pipeline;
 pub mod task;
 pub mod transform;
