+++ src/parser/mod.rs
@@ -1,4 +1,5 @@
 pub mod common;
+pub mod config;
 pub mod drop;
 pub mod join;
 pub mod model;
